LayoutTests/accessibility/ios-simulator/ios-search-predicate-expected.txt

start
first heading level 1

second heading level 1

heading level 2

heading level 3

heading level 4

heading level 5

heading level 6

This tests the ability to search for accessible elements by key or text.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS resultElement.childAtIndex(0).description is 'AXLabel: first heading level 1'
PASS resultElement.childAtIndex(0).description is 'AXLabel: heading level 2'
PASS resultElement.childAtIndex(0).description is 'AXLabel: heading level 3'
PASS resultElement.childAtIndex(0).description is 'AXLabel: heading level 4'
PASS resultElement.childAtIndex(0).description is 'AXLabel: heading level 5'
PASS resultElement.childAtIndex(0).description is 'AXLabel: heading level 6'
PASS successfullyParsed is true

TEST COMPLETE
